Move app/mirage to mirage
So that the mirage fixtures aren't included in the production build.

http://www.ember-cli-mirage.com/docs/v0.2.x/upgrading/#x--02-upgrade-guide
