--- app/mirage/scenarios/default.js
@@ -1,2 +1,0 @@
-export default function(/* server */) {
-}
