+++ mirage/fixtures/categories.js
